# docs/public-api/references/swagger/process/flowable-swagger-process.yaml
...
swagger: "2.0"
info:
  description: "# flowable / flow??b(??)l /\n\n- a compact and highly efficient workflow\
    \ and Business Process Management (BPM) platform for developers, system admins\
    \ and business users.\n- a lightning fast, tried and tested BPMN 2 process engine\
    \ written in Java. It is Apache 2.0 licensed open source, with a committed community.\n\
    - can run embedded in a Java application, or as a service on a server, a cluster,\
    \ and in the cloud. It integrates perfectly with Spring. With a rich Java and\
    \ REST API, it is the ideal engine for orchestrating human or system activities."
  version: "v1"
  title: "Flowable Process REST API"
  contact:
    name: "Flowable"
    url: "http://www.flowable.org/"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:8080"
basePath: "/flowable-rest/service"
...
schemes:
- "http"
- "https"
paths:
...
  /runtime/executions:
    get:
      tags:
      - "Executions"
      summary: "List of executions"
      description: ""
      operationId: "listExecutions"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "query"
        description: "Only return models with the given version."
        required: false
        type: "string"
      - name: "activityId"
        in: "query"
        description: "Only return executions with the given activity id."
        required: false
        type: "string"
      - name: "processDefinitionKey"
        in: "query"
        description: "Only return process instances with the given process definition\
          \ key."
        required: false
        type: "string"
      - name: "processDefinitionId"
        in: "query"
        description: "Only return process instances with the given process definition\
          \ id."
        required: false
        type: "string"
      - name: "processInstanceId"
        in: "query"
        description: "Only return executions which are part of the process instance\
          \ with the given id."
        required: false
        type: "string"
      - name: "messageEventSubscriptionName"
        in: "query"
        description: "Only return executions which are subscribed to a message with\
          \ the given name."
        required: false
        type: "string"
      - name: "signalEventSubscriptionName"
        in: "query"
        description: "Only return executions which are subscribed to a signal with\
          \ the given name."
        required: false
        type: "string"
      - name: "parentId"
        in: "query"
        description: "Only return executions which are a direct child of the given\
          \ execution."
        required: false
        type: "string"
      - name: "tenantId"
        in: "query"
        description: "Only return process instances with the given tenantId."
        required: false
        type: "string"
      - name: "tenantIdLike"
        in: "query"
        description: "Only return process instances with a tenantId like the given\
          \ value."
        required: false
        type: "string"
      - name: "withoutTenantId"
        in: "query"
        description: "If true, only returns process instances without a tenantId set.\
          \ If false, the withoutTenantId parameter is ignored."
        required: false
        type: "boolean"
      - name: "sort"
        in: "query"
        description: "Property to sort on, to be used together with the order."
        required: false
        type: "string"
        enum:
        - "processInstanceId"
        - "processDefinitionId"
        - "processDefinitionKey"
        - "tenantId"
      responses:
        200:
          description: "Indicates request was successful and the executions are returned"
          schema:
            $ref: "#/definitions/DataResponseExecutionResponse"
        400:
          description: "Indicates a parameter was passed in the wrong format . The\
            \ status-message contains additional information."
      security:
      - basicAuth: []
    put:
      tags:
      - "Executions"
      summary: "Signal event received"
      description: ""
      operationId: "executeExecutionAction"
      parameters:
      - in: "body"
        name: "body"
        required: false
        schema:
          $ref: "#/definitions/ExecutionActionRequest"
      responses:
        204:
          description: "Indicates request was successful and the executions are returned"
        404:
          description: "Indicates a parameter was passed in the wrong format . The\
            \ status-message contains additional information."
      security:
      - basicAuth: []
...
  /runtime/signals:
    post:
      tags:
      - "Runtime"
      summary: "Signal event received"
      description: ""
      operationId: "signalEventReceived"
      parameters:
      - in: "body"
        name: "body"
        required: false
        schema:
          $ref: "#/definitions/SignalEventReceivedRequest"
      responses:
        202:
          description: "Indicated signal processing is queued as a job, ready to be\
            \ executed."
        204:
          description: "Indicated signal has been processed and no errors occurred."
        400:
          description: "Signal not processed. The signal name is missing or variables\
            \ are used together with async, which is not allowed. Response body contains\
            \ additional information about the error."
      security:
      - basicAuth: []
...
  /runtime/tasks/{taskId}/identitylinks/{family}:
    get:
      tags:
      - "Task Identity Links"
      summary: "List identity links for a task for either groups or users"
      description: "Returns only identity links targetting either users or groups.\
        \ Response body and status-codes are exactly the same as when getting the\
        \ full list of identity links for a task."
      operationId: "listIdentityLinksForFamily"
      produces:
      - "application/json"
      parameters:
      - name: "taskId"
        in: "path"
        required: true
        type: "string"
      - name: "family"
        in: "path"
        required: true
        type: "string"
      responses:
        200:
          description: "Indicates the task was found and the requested identity links\
            \ are returned."
          schema:
            type: "array"
            items:
              $ref: "#/definitions/RestIdentityLink"
        404:
          description: "Indicates the requested task was not found."
      security:
      - basicAuth: []
...
securityDefinitions:
  basicAuth:
    type: "basic"
definitions:
...
  DataResponseExecutionResponse:
    type: "object"
    properties:
      data:
        type: "array"
        items:
          $ref: "#/definitions/ExecutionResponse"
      total:
        type: "integer"
        format: "int64"
      start:
        type: "integer"
        format: "int32"
      sort:
        type: "string"
      order:
        type: "string"
      size:
        type: "integer"
        format: "int32"
...
  ExecutionActionRequest:
    type: "object"
    required:
    - "action"
    properties:
      action:
        type: "string"
        example: "signalEventReceived"
        description: "Action to perform: Either signal, trigger, signalEventReceived\
          \ or messageEventReceived"
      signalName:
        type: "string"
        example: "My Signal"
        description: "Name of the signal"
      messageName:
        type: "string"
        example: "My Signal"
        description: "Message of the signal"
      variables:
        type: "array"
        items:
          $ref: "#/definitions/RestVariable"
      transientVariables:
        type: "array"
        items:
          $ref: "#/definitions/RestVariable"
...
  ExecutionResponse:
    type: "object"
    properties:
      id:
        type: "string"
        example: "5"
      url:
        type: "string"
        example: "http://localhost:8182/runtime/executions/5"
      parentId:
        type: "string"
        example: "null"
      parentUrl:
        type: "string"
        example: "null"
      superExecutionId:
        type: "string"
        example: "null"
      superExecutionUrl:
        type: "string"
        example: "null"
      processInstanceId:
        type: "string"
        example: "5"
      processInstanceUrl:
        type: "string"
        example: "http://localhost:8182/runtime/process-instances/5"
      suspended:
        type: "boolean"
      activityId:
        type: "string"
        example: "null"
      tenantId:
        type: "string"
        example: "null"
...
  RestIdentityLink:
    type: "object"
    properties:
      url:
        type: "string"
      user:
        type: "string"
        example: "kermit"
      group:
        type: "string"
        example: "sales"
      type:
        type: "string"
        example: "candidate"
  RestVariable:
    type: "object"
    properties:
      name:
        type: "string"
        example: "myVariable"
        title: "Name of the variable"
      type:
        type: "string"
        example: "string"
        title: "Type of the variable."
        description: "When writing a variable and this value is omitted, the type will be deducted from the raw JSON-attribute request type and is limited to either string, double, integer and boolean. It’s advised to always include a type to make sure no wrong assumption about the type can be done."
      value:
        type: "object"
        example: "test"
        title: "Value of the variable."
        description: "When writing a variable and value is omitted, null will be used as value."
      valueUrl:
        type: "string"
        example: "http://...."
        title: "Url value to the binary"
        description: "When reading a variable of type binary or serializable, this attribute will point to the URL where the raw binary data can be fetched from."
      scope:
        title: "scope of the variable."
        type: "string"
        example: "global"
        description: "If local, the variable is explicitly defined on the resource it’s requested from. When global, the variable is defined on the parent (or any parent in the parent-tree) of the resource it’s requested from. When writing a variable and the scope is omitted, global is assumed."
...
  SignalEventReceivedRequest:
    type: "object"
    properties:
      signalName:
        type: "string"
        description: "Name of the signal"
      variables:
        type: "array"
        description: "Array of variables (in the general variables format) to use\
          \ as payload to pass along with the signal. Cannot be used in case async\
          \ is set to true, this will result in an error."
        items:
          $ref: "#/definitions/RestVariable"
      tenantId:
        type: "string"
        description: "ID of the tenant that the signal event should be processed in"
      async:
        type: "boolean"
        description: "If true, handling of the signal will happen asynchronously.\
          \ Return code will be 202 - Accepted to indicate the request is accepted\
          \ but not yet executed. If false,\n                    handling the signal\
          \ will be done immediately and result (200 - OK) will only return after\
          \ this completed successfully. Defaults to false if omitted."
...